@@ -14,54 +14,32 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Session;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using Org.BouncyCastle.Crypto.Generators;
 using Seterator.Models;
+using Seterator.Services;
 using static Seterator.Services.SessionExtensions;
 
 namespace Seterator.Controllers
 {
     public class AccountController : Controller
     {
-        private readonly DatabaseContext database;
-        private static readonly byte[] salt = new byte[]
+        private readonly AccountService account;
+        private readonly AuthService auth;
+        
+        public AccountController(AccountService account, AuthService auth)
         {
-            0xad, 0x34, 0xff, 0x45, 0xf2, 0x12, 0x34, 0xfa,
-            0xad, 0x34, 0xff, 0x45, 0xf2, 0x12, 0x34, 0xfa
-        };
-        public AccountController(DatabaseContext database)
-        {
-            this.database = database;
+            this.account = account;
+            this.auth = auth;
         }
-
+        
+        public IActionResult Main()
+        {
+            return View();
+        }
+        
         [HttpGet]
         public IActionResult Login()
         {
             return View();
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult Login(ViewModels.LoginModel model)
-        {
-            Contract.Assert(model != null);
-            if (ModelState.IsValid)
-            {
-                Models.User user = database.Users.Include(x => x.Roles).FirstOrDefault(u => u.Login == model.Login);
-                if (user == null || !BCrypt.Generate(Encoding.Unicode.GetBytes(model.Password), salt, 8).SequenceEqual(user.PasswordHash))
-                {
-                    ModelState.AddModelError("", "Неверный логин или пароль");
-                }
-                else
-                {
-                    Authenticate(user).ConfigureAwait(false);
-                    return RedirectToAction("Index", "Home");
-                }
-            }
-            else
-            {
-                ModelState.AddModelError("", "Неправильно введены данные");
-            }
-            return View(model);
         }
 
         [HttpGet]
@@ -70,61 +48,46 @@ namespace Seterator.Controllers
             return View();
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Register(ViewModels.RegistrationModel data)
-        {
-            var user = database.Users.FirstOrDefault(u => u.Login == data.Login);
-            if (data.Password == null)
-                data.Password = "password";
-            if (user == null)
-            {
-                var newUser = new Models.User()
-                {
-                    Login = data.Login,
-                    PasswordHash = BCrypt.Generate(Encoding.Unicode.GetBytes(data.Password), salt, 8)
-                };
-                database.Users.Add(newUser);
-                await database.SaveChangesAsync().ConfigureAwait(true);
-                _ = Authenticate(newUser);
-                return RedirectToAction("Index", "Home");
-            }  
-            else
-            {
-                ModelState.AddModelError("", "Некорректные данные");
-                return View(data);
-            }
-        }
-
-        public IActionResult Main()
-        {
-            return View();
-        }
-
-        private async Task Authenticate(User user)
-        {
-            var claims = new List<Claim>();
-            claims.Add(new Claim(ClaimsIdentity.DefaultNameClaimType, user.Login));
-            foreach (var role in user.Roles)
-            {
-                claims.Add(new Claim(ClaimsIdentity.DefaultRoleClaimType, role.UserRole.ToString()));
-            }
-            ClaimsIdentity id = new ClaimsIdentity(claims:             claims,
-                                                   authenticationType: "ApplicationCookie",
-                                                   nameType:           ClaimsIdentity.DefaultNameClaimType,
-                                                   roleType:           ClaimsIdentity.DefaultRoleClaimType);
-            await HttpContext
-                .SignInAsync(CookieAuthenticationDefaults.AuthenticationScheme, 
-                             new ClaimsPrincipal(id))
-                .ConfigureAwait(false);
-        }
-
         public async Task<IActionResult> Logout()
         {
-            await HttpContext
-                .SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme)
-                .ConfigureAwait(false);
+            await auth.Logout();
             return RedirectToAction("Index", "Home");
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Login(ViewModels.LoginModel model)
+        {
+        
+            var canLogin = account.TryLogin(model.Login, model.Password);
+            if (canLogin)
+            {
+                await auth.Authenticate(model.Login, Enumerable.Empty<string>());
+                return RedirectToAction("Index", "Home");
+            }
+            else
+            {
+                ModelState.AddModelError("", "Неверный логин или пароль");
+            }
+            return View(model);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Register(ViewModels.RegistrationModel model)
+        {
+            var loginAvailable = account.IsLoginAvailable(model.Login);
+            if (loginAvailable)
+            {
+                await account.Register(model.Login, model.Password);
+                await auth.Authenticate(model.Login, Enumerable.Empty<string>());
+                return RedirectToAction("Index", "Home");
+            }
+            else
+            {
+                ModelState.AddModelError("", "Пользователь с таким именем уже существует");
+                return View(model);
+            }
         }
     }
 }
