@@ -7,12 +7,8 @@ namespace Seterator.ViewModels
         [Required(ErrorMessage = "Не указан логин")]
         public string Login { get; set; }
 
+        [DataType(DataType.Password)]
         [Required(ErrorMessage = "Не указан пароль")]
-        [DataType(DataType.Password)]
         public string Password { get; set; }
-
-        [DataType(DataType.Password)]
-        [Compare("Password", ErrorMessage = "Пароль введен неверно")]
-        public string PasswordConfirmation { get; set; }
     }
 }
